pyproject.toml
[project]
name = "vktestbackendmlengineer"
version = "0.1.0"
description = "VK chat bot"
authors = [
    {name = "Alexey",email = "rybolovlevalexey@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "vk-api (>=11.9.9,<12.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
